pair-wise Spearman rank co-occurrence, written in efficient c++ code. Function from the phylosmith-package.

A rewrite of the pair-wise Spearman rank co-occurrence routine written by Jin Choi. The routine has been adapted to integrate with the Rcpp-package API.

co_occurrence(phyloseq_obj, treatment = NULL, subset = NULL, rho = 0, p = 0.05, cores = 0)

Arguments

phyloseq_obj
A phyloseq-class object.

treatment
Column name as a string or numeric in the sample_data. This can be a vector of multiple columns and they will be combined into a new column.

subset
A factor within the treatment. This will remove any samples that to not contain this factor. This can be a vector of multiple factors to subset on.

rho
numeric The rho-value cutoff. All returned co-occurrences will have a rho-value less than or equal to rho or less than or equal to -rho.

p
numeric The p-value cutoff. All returned co-occurrences will have a p-value less than or equal to p.

cores
numeric Number of CPU cores to use for the pair-wise permutations. Default (0) uses max cores available. Parallelization not available for systems running MacOS without openMP configuration.

Value

data.table

See also

Examples

co_occurrence(soil_column, treatment = c('Matrix', 'Treatment'),
subset = 'Amended', rho = 0.8, p = 0.05, cores = 0)

#>              Treatment   OTU_1   OTU_2       rho            p
#>    1:     Soil Amended   OTU_5   OTU_6 0.8188239 4.183781e-34
#>    2:     Soil Amended   OTU_5  OTU_72 0.8066509 2.111979e-32
#>    3:     Soil Amended   OTU_5  OTU_74 0.8449522 3.102246e-38
#>    4:     Soil Amended   OTU_5  OTU_86 0.8111768 5.080468e-33
#>    5:     Soil Amended   OTU_5 OTU_197 0.8203081 2.542049e-34
#>   ---                                                        
#> 1122: Effluent Amended OTU_536 OTU_919 0.8046293 4.381430e-22
#> 1123: Effluent Amended OTU_544 OTU_615 0.8209210 1.279868e-23
#> 1124: Effluent Amended OTU_544  OTU_94 0.8209706 1.265497e-23
#> 1125: Effluent Amended OTU_544 OTU_315 0.8209706 1.265497e-23
#> 1126: Effluent Amended OTU_544 OTU_317 0.8209706 1.265497e-23