Find taxa shared between treatments of a phyloseq object. Function from the phylosmith-package.

Inputs a phyloseq-class object and finds which taxa are shared between all of the specified treatments from data in the sample_data()), or every sample in the dataset.

common_taxa(phyloseq_obj, treatment = NULL, subset = NULL, n = 'all')

Arguments

phyloseq_obj
A phyloseq-class object. It must contain sample_data()) with information about each sample, and it must contain tax_table()) with information about each taxa/gene.

treatment
Column name as a string or numeric in the sample_data. This can be a vector of multiple columns and they will be combined into a new column.

subset
A factor within the treatment. This will remove any samples that to not contain this factor. This can be a vector of multiple factors to subset on.

n
Number of treatment groups that need to share the taxa to be considered a common taxa.

Value

vector

See also

Examples

common_taxa(soil_column, treatment = 'Treatment',
subset = 'Amended', n = 'all')

#>   [1] "OTU_1"   "OTU_10"  "OTU_100" "OTU_101" "OTU_102" "OTU_103" "OTU_104"
#>   [8] "OTU_105" "OTU_106" "OTU_108" "OTU_109" "OTU_11"  "OTU_110" "OTU_111"
#>  [15] "OTU_112" "OTU_114" "OTU_115" "OTU_116" "OTU_117" "OTU_118" "OTU_119"
#>  [22] "OTU_12"  "OTU_120" "OTU_121" "OTU_122" "OTU_123" "OTU_124" "OTU_125"
#>  [29] "OTU_126" "OTU_127" "OTU_128" "OTU_129" "OTU_13"  "OTU_130" "OTU_131"
#>  [36] "OTU_132" "OTU_133" "OTU_134" "OTU_135" "OTU_136" "OTU_137" "OTU_138"
#>  [43] "OTU_139" "OTU_14"  "OTU_140" "OTU_141" "OTU_142" "OTU_143" "OTU_144"
#>  [50] "OTU_145" "OTU_146" "OTU_147" "OTU_148" "OTU_149" "OTU_15"  "OTU_150"
#>  [57] "OTU_151" "OTU_152" "OTU_153" "OTU_154" "OTU_155" "OTU_156" "OTU_157"
#>  [64] "OTU_158" "OTU_159" "OTU_16"  "OTU_160" "OTU_161" "OTU_162" "OTU_163"
#>  [71] "OTU_164" "OTU_165" "OTU_166" "OTU_167" "OTU_169" "OTU_17"  "OTU_170"
#>  [78] "OTU_171" "OTU_172" "OTU_173" "OTU_174" "OTU_175" "OTU_176" "OTU_177"
#>  [85] "OTU_178" "OTU_179" "OTU_18"  "OTU_180" "OTU_181" "OTU_182" "OTU_183"
#>  [92] "OTU_184" "OTU_185" "OTU_186" "OTU_187" "OTU_188" "OTU_189" "OTU_19" 
#>  [99] "OTU_190" "OTU_191" "OTU_192" "OTU_193" "OTU_194" "OTU_195" "OTU_197"
#> [106] "OTU_198" "OTU_199" "OTU_2"   "OTU_20"  "OTU_200" "OTU_201" "OTU_202"
#> [113] "OTU_203" "OTU_204" "OTU_206" "OTU_208" "OTU_209" "OTU_21"  "OTU_210"
#> [120] "OTU_211" "OTU_212" "OTU_215" "OTU_216" "OTU_217" "OTU_218" "OTU_219"
#> [127] "OTU_22"  "OTU_220" "OTU_221" "OTU_222" "OTU_223" "OTU_224" "OTU_225"
#> [134] "OTU_226" "OTU_227" "OTU_228" "OTU_23"  "OTU_230" "OTU_231" "OTU_232"
#> [141] "OTU_233" "OTU_234" "OTU_235" "OTU_236" "OTU_237" "OTU_238" "OTU_239"
#> [148] "OTU_24"  "OTU_240" "OTU_241" "OTU_242" "OTU_243" "OTU_244" "OTU_245"
#> [155] "OTU_246" "OTU_247" "OTU_249" "OTU_25"  "OTU_251" "OTU_252" "OTU_254"
#> [162] "OTU_255" "OTU_256" "OTU_257" "OTU_258" "OTU_259" "OTU_26"  "OTU_260"
#> [169] "OTU_261" "OTU_262" "OTU_263" "OTU_264" "OTU_265" "OTU_266" "OTU_267"
#> [176] "OTU_268" "OTU_269" "OTU_27"  "OTU_270" "OTU_271" "OTU_272" "OTU_273"
#> [183] "OTU_274" "OTU_275" "OTU_276" "OTU_277" "OTU_278" "OTU_279" "OTU_28" 
#> [190] "OTU_280" "OTU_281" "OTU_282" "OTU_283" "OTU_284" "OTU_285" "OTU_286"
#> [197] "OTU_287" "OTU_288" "OTU_289" "OTU_29"  "OTU_290" "OTU_291" "OTU_292"
#> [204] "OTU_293" "OTU_294" "OTU_295" "OTU_296" "OTU_297" "OTU_298" "OTU_30" 
#> [211] "OTU_301" "OTU_302" "OTU_303" "OTU_304" "OTU_305" "OTU_306" "OTU_307"
#> [218] "OTU_308" "OTU_309" "OTU_31"  "OTU_310" "OTU_311" "OTU_312" "OTU_313"
#> [225] "OTU_314" "OTU_315" "OTU_316" "OTU_317" "OTU_318" "OTU_319" "OTU_32" 
#> [232] "OTU_320" "OTU_321" "OTU_322" "OTU_323" "OTU_324" "OTU_325" "OTU_326"
#> [239] "OTU_327" "OTU_329" "OTU_33"  "OTU_330" "OTU_331" "OTU_332" "OTU_334"
#> [246] "OTU_335" "OTU_336" "OTU_34"  "OTU_341" "OTU_342" "OTU_343" "OTU_344"
#> [253] "OTU_345" "OTU_346" "OTU_347" "OTU_349" "OTU_35"  "OTU_350" "OTU_351"
#> [260] "OTU_352" "OTU_353" "OTU_355" "OTU_356" "OTU_357" "OTU_358" "OTU_359"
#> [267] "OTU_36"  "OTU_360" "OTU_361" "OTU_362" "OTU_363" "OTU_364" "OTU_365"
#> [274] "OTU_366" "OTU_367" "OTU_368" "OTU_369" "OTU_37"  "OTU_370" "OTU_371"
#> [281] "OTU_372" "OTU_373" "OTU_374" "OTU_375" "OTU_376" "OTU_378" "OTU_379"
#> [288] "OTU_38"  "OTU_380" "OTU_381" "OTU_384" "OTU_385" "OTU_386" "OTU_387"
#> [295] "OTU_388" "OTU_389" "OTU_39"  "OTU_390" "OTU_391" "OTU_393" "OTU_394"
#> [302] "OTU_396" "OTU_397" "OTU_399" "OTU_4"   "OTU_40"  "OTU_400" "OTU_402"
#> [309] "OTU_403" "OTU_404" "OTU_405" "OTU_406" "OTU_407" "OTU_408" "OTU_409"
#> [316] "OTU_41"  "OTU_410" "OTU_411" "OTU_412" "OTU_413" "OTU_414" "OTU_415"
#> [323] "OTU_416" "OTU_417" "OTU_418" "OTU_419" "OTU_42"  "OTU_421" "OTU_422"
#> [330] "OTU_423" "OTU_424" "OTU_425" "OTU_426" "OTU_428" "OTU_43"  "OTU_430"
#> [337] "OTU_431" "OTU_432" "OTU_433" "OTU_434" "OTU_435" "OTU_436" "OTU_437"
#> [344] "OTU_438" "OTU_439" "OTU_44"  "OTU_440" "OTU_441" "OTU_442" "OTU_443"
#> [351] "OTU_444" "OTU_445" "OTU_448" "OTU_449" "OTU_45"  "OTU_450" "OTU_452"
#> [358] "OTU_453" "OTU_454" "OTU_455" "OTU_456" "OTU_457" "OTU_458" "OTU_459"
#> [365] "OTU_46"  "OTU_460" "OTU_461" "OTU_462" "OTU_463" "OTU_464" "OTU_465"
#> [372] "OTU_466" "OTU_467" "OTU_468" "OTU_469" "OTU_47"  "OTU_470" "OTU_471"
#> [379] "OTU_472" "OTU_473" "OTU_474" "OTU_475" "OTU_476" "OTU_478" "OTU_48" 
#> [386] "OTU_480" "OTU_481" "OTU_482" "OTU_483" "OTU_484" "OTU_485" "OTU_486"
#> [393] "OTU_487" "OTU_488" "OTU_489" "OTU_49"  "OTU_490" "OTU_491" "OTU_492"
#> [400] "OTU_493" "OTU_494" "OTU_495" "OTU_496" "OTU_497" "OTU_498" "OTU_499"
#> [407] "OTU_5"   "OTU_50"  "OTU_502" "OTU_503" "OTU_504" "OTU_505" "OTU_506"
#> [414] "OTU_507" "OTU_508" "OTU_509" "OTU_51"  "OTU_510" "OTU_511" "OTU_512"
#> [421] "OTU_513" "OTU_514" "OTU_515" "OTU_516" "OTU_517" "OTU_518" "OTU_519"
#> [428] "OTU_52"  "OTU_520" "OTU_521" "OTU_522" "OTU_523" "OTU_524" "OTU_525"
#> [435] "OTU_526" "OTU_527" "OTU_528" "OTU_529" "OTU_53"  "OTU_530" "OTU_531"
#> [442] "OTU_532" "OTU_533" "OTU_534" "OTU_535" "OTU_536" "OTU_537" "OTU_538"
#> [449] "OTU_539" "OTU_54"  "OTU_540" "OTU_541" "OTU_543" "OTU_544" "OTU_545"
#> [456] "OTU_546" "OTU_547" "OTU_548" "OTU_549" "OTU_55"  "OTU_550" "OTU_551"
#> [463] "OTU_552" "OTU_553" "OTU_554" "OTU_555" "OTU_556" "OTU_557" "OTU_558"
#> [470] "OTU_559" "OTU_56"  "OTU_560" "OTU_561" "OTU_562" "OTU_563" "OTU_564"
#> [477] "OTU_565" "OTU_566" "OTU_568" "OTU_569" "OTU_57"  "OTU_570" "OTU_571"
#> [484] "OTU_572" "OTU_573" "OTU_574" "OTU_575" "OTU_577" "OTU_578" "OTU_579"
#> [491] "OTU_58"  "OTU_580" "OTU_581" "OTU_582" "OTU_583" "OTU_584" "OTU_585"
#> [498] "OTU_586" "OTU_587" "OTU_588" "OTU_589" "OTU_59"  "OTU_590" "OTU_591"
#> [505] "OTU_592" "OTU_593" "OTU_594" "OTU_595" "OTU_596" "OTU_597" "OTU_598"
#> [512] "OTU_599" "OTU_6"   "OTU_60"  "OTU_600" "OTU_601" "OTU_602" "OTU_603"
#> [519] "OTU_604" "OTU_605" "OTU_606" "OTU_607" "OTU_608" "OTU_609" "OTU_61" 
#> [526] "OTU_610" "OTU_611" "OTU_612" "OTU_613" "OTU_614" "OTU_615" "OTU_616"
#> [533] "OTU_617" "OTU_618" "OTU_619" "OTU_62"  "OTU_620" "OTU_621" "OTU_623"
#> [540] "OTU_624" "OTU_625" "OTU_627" "OTU_628" "OTU_629" "OTU_63"  "OTU_631"
#> [547] "OTU_632" "OTU_633" "OTU_635" "OTU_636" "OTU_637" "OTU_638" "OTU_639"
#> [554] "OTU_64"  "OTU_640" "OTU_641" "OTU_642" "OTU_643" "OTU_644" "OTU_645"
#> [561] "OTU_646" "OTU_647" "OTU_648" "OTU_649" "OTU_65"  "OTU_650" "OTU_651"
#> [568] "OTU_652" "OTU_653" "OTU_654" "OTU_655" "OTU_656" "OTU_657" "OTU_658"
#> [575] "OTU_659" "OTU_66"  "OTU_660" "OTU_661" "OTU_662" "OTU_663" "OTU_664"
#> [582] "OTU_665" "OTU_666" "OTU_667" "OTU_668" "OTU_669" "OTU_67"  "OTU_670"
#> [589] "OTU_672" "OTU_673" "OTU_675" "OTU_676" "OTU_677" "OTU_678" "OTU_679"
#> [596] "OTU_68"  "OTU_680" "OTU_681" "OTU_683" "OTU_684" "OTU_685" "OTU_686"
#> [603] "OTU_688" "OTU_689" "OTU_69"  "OTU_690" "OTU_691" "OTU_692" "OTU_694"
#> [610] "OTU_696" "OTU_697" "OTU_699" "OTU_7"   "OTU_70"  "OTU_700" "OTU_701"
#> [617] "OTU_702" "OTU_703" "OTU_704" "OTU_706" "OTU_71"  "OTU_710" "OTU_711"
#> [624] "OTU_712" "OTU_713" "OTU_714" "OTU_715" "OTU_716" "OTU_717" "OTU_718"
#> [631] "OTU_719" "OTU_72"  "OTU_720" "OTU_721" "OTU_722" "OTU_723" "OTU_724"
#> [638] "OTU_725" "OTU_726" "OTU_727" "OTU_728" "OTU_729" "OTU_73"  "OTU_730"
#> [645] "OTU_732" "OTU_733" "OTU_734" "OTU_735" "OTU_736" "OTU_738" "OTU_739"
#> [652] "OTU_74"  "OTU_740" "OTU_741" "OTU_743" "OTU_744" "OTU_746" "OTU_747"
#> [659] "OTU_748" "OTU_749" "OTU_75"  "OTU_750" "OTU_751" "OTU_752" "OTU_753"
#> [666] "OTU_754" "OTU_755" "OTU_756" "OTU_757" "OTU_758" "OTU_759" "OTU_76" 
#> [673] "OTU_760" "OTU_761" "OTU_762" "OTU_763" "OTU_764" "OTU_765" "OTU_766"
#> [680] "OTU_767" "OTU_768" "OTU_769" "OTU_77"  "OTU_770" "OTU_771" "OTU_772"
#> [687] "OTU_773" "OTU_774" "OTU_775" "OTU_776" "OTU_777" "OTU_778" "OTU_78" 
#> [694] "OTU_780" "OTU_781" "OTU_782" "OTU_784" "OTU_785" "OTU_786" "OTU_787"
#> [701] "OTU_788" "OTU_789" "OTU_79"  "OTU_790" "OTU_791" "OTU_792" "OTU_793"
#> [708] "OTU_794" "OTU_795" "OTU_796" "OTU_798" "OTU_799" "OTU_8"   "OTU_80" 
#> [715] "OTU_800" "OTU_801" "OTU_802" "OTU_805" "OTU_806" "OTU_807" "OTU_808"
#> [722] "OTU_809" "OTU_81"  "OTU_810" "OTU_811" "OTU_813" "OTU_814" "OTU_815"
#> [729] "OTU_816" "OTU_817" "OTU_818" "OTU_82"  "OTU_820" "OTU_822" "OTU_823"
#> [736] "OTU_825" "OTU_826" "OTU_827" "OTU_83"  "OTU_830" "OTU_831" "OTU_833"
#> [743] "OTU_834" "OTU_836" "OTU_837" "OTU_838" "OTU_839" "OTU_84"  "OTU_840"
#> [750] "OTU_841" "OTU_842" "OTU_843" "OTU_844" "OTU_845" "OTU_846" "OTU_847"
#> [757] "OTU_848" "OTU_849" "OTU_85"  "OTU_850" "OTU_851" "OTU_852" "OTU_853"
#> [764] "OTU_854" "OTU_855" "OTU_856" "OTU_857" "OTU_858" "OTU_859" "OTU_86" 
#> [771] "OTU_860" "OTU_862" "OTU_863" "OTU_864" "OTU_865" "OTU_866" "OTU_867"
#> [778] "OTU_868" "OTU_869" "OTU_87"  "OTU_870" "OTU_872" "OTU_873" "OTU_874"
#> [785] "OTU_875" "OTU_876" "OTU_877" "OTU_878" "OTU_879" "OTU_88"  "OTU_880"
#> [792] "OTU_881" "OTU_883" "OTU_884" "OTU_885" "OTU_886" "OTU_887" "OTU_888"
#> [799] "OTU_889" "OTU_89"  "OTU_891" "OTU_892" "OTU_893" "OTU_894" "OTU_895"
#> [806] "OTU_896" "OTU_897" "OTU_898" "OTU_899" "OTU_9"   "OTU_90"  "OTU_900"
#> [813] "OTU_901" "OTU_902" "OTU_903" "OTU_904" "OTU_905" "OTU_906" "OTU_907"
#> [820] "OTU_908" "OTU_909" "OTU_91"  "OTU_910" "OTU_911" "OTU_912" "OTU_913"
#> [827] "OTU_914" "OTU_915" "OTU_916" "OTU_917" "OTU_919" "OTU_92"  "OTU_920"
#> [834] "OTU_93"  "OTU_94"  "OTU_95"  "OTU_96"  "OTU_97"  "OTU_98"  "OTU_99"