Merge samples based on common factor within sample_data. Function from the phylosmith-package.

Inputs a phyloseq object and merges the samples that meet the specified criteria into a single sample. This is meant for replicates, or samples statistically proven to not be significantly different. This should be used with caution as it may be a misleading representation of the data.

conglomerate_samples(phyloseq_obj, treatment, subset = NULL,
merge_on = treatment)

Arguments

phyloseq_obj
A phyloseq-class object. It must contain sample_data()) with information about each sample, and it must contain tax_table()) with information about each taxa/gene.

treatment
Column name as a string or numeric in the sample_data. This can be a vector of multiple columns and they will be combined into a new column.

subset
A factor within the treatment. This will remove any samples that to not contain this factor. This can be a vector of multiple factors to subset on.

merge_on
Defines which variable the data is merged according to. This needs to be a column name as a string or numeric in the sample_data. This can be a vector of multiple columns and they will be combined into a new column.

Value

phyloseq-object

See also

Examples

conglomerate_samples(soil_column, treatment = c('Matrix', 'Treatment'), merge_on = 'Day')

#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 890 taxa and 21 samples ]
#> sample_data() Sample Data:       [ 21 samples by 4 sample variables ]
#> tax_table()   Taxonomy Table:    [ 890 taxa by 6 taxonomic ranks ]