Curate co-occurrence data. Function from the phylosmith-package.

Used to curate a co-occurrence table from the co_occurrence function. Takes a list of taxa and finds all pairs containing any of those taxa.

curate_co_occurrence(co_occurrence_table, taxa_of_interest,
number_of_treatments = 1)

Arguments

co_occurrence_table
Table of the co-occurrence of taxa/genes in the phyloseq_obj, computed using co_occurrence.

taxa_of_interest
A list or vector of taxa names, like those generated with unique_taxa.

number_of_treatments
How many treatments should the taxa of interest be seen in? Requires integer or 'all' (default = 1).

Value

data.table

See also