Melt a phyloseq object into a data.table. Function from the phylosmith-package.

melt_phyloseq inputs a phyloseq object and melts its otu_table, taxa_tale, and sample_Data into a single into a data.table.

melt_phyloseq(phyloseq_obj)

Arguments

phyloseq_obj
A phyloseq-class object. It must contain sample_data() with information about each sample, and it must contain tax_table()) with information about each taxa/gene.

Value

data.table

See also

Examples

melt_phyloseq(soil_column)

#>             OTU Sample Abundance   Matrix Treatment Day   Domain
#>      1: OTU_697   S_88     14648 Effluent Unamended  38 Bacteria
#>      2: OTU_676   S_85     12559 Effluent Unamended  38 Bacteria
#>      3: OTU_911   S_38     12076 Effluent   Amended  10 Bacteria
#>      4: OTU_697   S_94      8733 Effluent Unamended  38 Bacteria
#>      5: OTU_911   S_37      7479 Effluent   Amended  10 Bacteria
#>     ---                                                         
#> 442516:  OTU_99   S_95         0 Effluent   Amended  38 Bacteria
#> 442517:  OTU_99   S_96         0 Effluent   Amended  38 Bacteria
#> 442518:  OTU_99   S_97         0 Effluent   Amended  38 Bacteria
#> 442519:  OTU_99   S_98         0 Effluent   Amended  38 Bacteria
#> 442520:  OTU_99   S_99         0 Effluent   Amended  38 Bacteria
#>                  Phylum              Class              Order
#>      1:  Proteobacteria Betaproteobacteria    Burkholderiales
#>      2:  Proteobacteria Betaproteobacteria    Burkholderiales
#>      3: Verrucomicrobia   Verrucomicrobiae Verrucomicrobiales
#>      4:  Proteobacteria Betaproteobacteria    Burkholderiales
#>      5: Verrucomicrobia   Verrucomicrobiae Verrucomicrobiales
#>     ---                                                      
#> 442516:  Actinobacteria     Actinobacteria    Actinomycetales
#> 442517:  Actinobacteria     Actinobacteria    Actinomycetales
#> 442518:  Actinobacteria     Actinobacteria    Actinomycetales
#> 442519:  Actinobacteria     Actinobacteria    Actinomycetales
#> 442520:  Actinobacteria     Actinobacteria    Actinomycetales
#>                      Family                         Genus
#>      1:    Oxalobacteraceae unclassified_Oxalobacteraceae
#>      2:      Comamonadaceae                   Curvibacter
#>      3: Verrucomicrobiaceae               Prosthecobacter
#>      4:    Oxalobacteraceae unclassified_Oxalobacteraceae
#>      5: Verrucomicrobiaceae               Prosthecobacter
#>     ---                                                  
#> 442516:     Kineosporiaceae                   Kineosporia
#> 442517:     Kineosporiaceae                   Kineosporia
#> 442518:     Kineosporiaceae                   Kineosporia
#> 442519:     Kineosporiaceae                   Kineosporia
#> 442520:     Kineosporiaceae                   Kineosporia