Create an layout_igraph object of the co-occurrence from a phyloseq object. Function from the phylosmith-package.

Create an layout_igraph object of the co-occurrence from a phyloseq object. This can be input into the network_ps function, or used for other network creating scripts. The purpose is to be able to create reproducible and comparable graphics.

network_layout_ps(phyloseq_obj, classification = NULL,
treatment = NULL, subset = NULL, co_occurrence_table = NULL,
algorithm = 'fr')

Arguments

phyloseq_obj
A phyloseq-class object. It must contain sample_data()) with information about each sample, and it must contain tax_table()) with information about each taxa/gene.

classification
Column name as a string or number in the tax_table for the factor to use for node colors.

treatment
Column name as a string or number in the sample_data. This can be a vector of multiple columns and they will be combined into a new column.

subset
A factor within the treatment. This will remove any samples that to not contain this factor. This can be a vector of multiple factors to subset on.

co_occurrence_table
Table of the co-occurrence of taxa/genes in the phyloseq_obj, computed using co_occurrence. If no table is given, it will be computed with the phyloseq_obj, using the given treatment and p = 0.05.

algorithm
Supported igraph::layout_() algorithm.

Value

layout_igraph object

Examples

network_layout_ps(soil_column, treatment = c('Matrix', 'Treatment'),
subset = 'Soil Amended', algorithm = 'kk')

#>               x            y    name ggraph.orig_index circular ggraph.index
#> 1    -2.9644490   6.15810325   OTU_5                 1    FALSE            1
#> 2    -5.3779199   5.26897815   OTU_6                 2    FALSE            2
#> 3    -7.1089580   0.70149042   OTU_8                 3    FALSE            3
#> 4     4.5396908  -7.17076578  OTU_18                 4    FALSE            4
#> 5     5.4371053  -9.39704793  OTU_33                 5    FALSE            5
#> 6     7.5920108  -5.92674174  OTU_38                 6    FALSE            6
#> 7     6.7205701  -7.99630210  OTU_39                 7    FALSE            7
#> 8     5.9641745 -11.14080797  OTU_43                 8    FALSE            8
#> 9     3.7828585  -9.14451340  OTU_45                 9    FALSE            9
#> 10    5.5912066  -6.71493078  OTU_46                10    FALSE           10
#> 11   -1.9656612 -12.74887682  OTU_48                11    FALSE           11
#> 12    6.7687667   0.34600890  OTU_51                12    FALSE           12
#> 13    6.9453638  -0.99536734  OTU_52                13    FALSE           13
#> 14    8.3227404  -6.43691996  OTU_53                14    FALSE           14
#> 15    5.7899098  -2.84465687  OTU_55                15    FALSE           15
#> 16    8.0202737  -3.46646047  OTU_56                16    FALSE           16
#> 17    6.9054495 -10.04129479  OTU_57                17    FALSE           17
#> 18    4.6388473  -9.56573810  OTU_61                18    FALSE           18
#> 19    3.7910228  -7.86192573  OTU_66                19    FALSE           19
#> 20    5.8722122  -8.79176985  OTU_70                20    FALSE           20
#> 21   -2.7373857   4.73741163  OTU_72                21    FALSE           21
#> 22    1.9825058 -13.31049292  OTU_73                22    FALSE           22
#> 23   -5.0010682   4.70573775  OTU_74                23    FALSE           23
#> 24   -5.3029023   3.50871780  OTU_79                24    FALSE           24
#> 25   10.1881637  -6.22510858  OTU_82                25    FALSE           25
#> 26   -5.2946015   4.26988926  OTU_86                26    FALSE           26
#> 27    6.5207898   4.55761642  OTU_87                27    FALSE           27
#> 28    5.6562330   6.38585742  OTU_89                28    FALSE           28
#> 29   -1.8527678   5.13704539 OTU_108                29    FALSE           29
#> 30    7.8445209   6.00530557 OTU_134                30    FALSE           30
#> 31    8.1563244   4.05484861 OTU_144                31    FALSE           31
#> 32    2.9724690  -5.09254225 OTU_146                32    FALSE           32
#> 33   -3.0318909 -10.15656165 OTU_160                33    FALSE           33
#> 34   -1.2309099 -11.85132165 OTU_164                34    FALSE           34
#> 35    5.6268379  -6.06426065 OTU_166                35    FALSE           35
#> 36    0.1112746   4.76066202 OTU_174                36    FALSE           36
#> 37   -8.4077926  -8.78935466 OTU_176                37    FALSE           37
#> 38    5.6085613  -7.50650091 OTU_177                38    FALSE           38
#> 39   -8.5133179  -7.37207082 OTU_179                39    FALSE           39
#> 40    9.8968520   4.03940966 OTU_181                40    FALSE           40
#> 41    4.3880079  -8.84369541 OTU_183                41    FALSE           41
#> 42    8.6104760  -4.20203386 OTU_187                42    FALSE           42
#> 43    8.3305615  -8.36856528 OTU_190                43    FALSE           43
#> 44   -0.2409256   3.36851996 OTU_194                44    FALSE           44
#> 45   -3.5707892   4.17798131 OTU_197                45    FALSE           45
#> 46   -3.7711425   4.70711713 OTU_198                46    FALSE           46
#> 47   -3.8531412   5.38865643 OTU_204                47    FALSE           47
#> 48   -0.1794717   5.37617667 OTU_206                48    FALSE           48
#> 49   -4.8216510   5.48120812 OTU_208                49    FALSE           49
#> 50  -11.5575447  -7.33809732 OTU_209                50    FALSE           50
#> 51   -3.2520918  -0.22707876 OTU_212                51    FALSE           51
#> 52   -0.1206443   2.25467275 OTU_217                52    FALSE           52
#> 53   -5.5849506   2.03397063 OTU_220                53    FALSE           53
#> 54   -6.0607874   5.44089022 OTU_255                54    FALSE           54
#> 55   -6.6676184   4.71394388 OTU_263                55    FALSE           55
#> 56    6.5747083   6.84808522 OTU_283                56    FALSE           56
#> 57    8.5833097  -9.14552244 OTU_308                57    FALSE           57
#> 58    4.5184913  -4.10407534 OTU_309                58    FALSE           58
#> 59    6.3538959  -7.65597554 OTU_310                59    FALSE           59
#> 60   -5.0668655  -8.46594537 OTU_321                60    FALSE           60
#> 61    0.6757017 -13.19003105 OTU_327                61    FALSE           61
#> 62   -0.8659517 -13.82881947 OTU_329                62    FALSE           62
#> 63    6.5322593  -5.00350238 OTU_330                63    FALSE           63
#> 64    2.2726126   4.75705201 OTU_331                64    FALSE           64
#> 65   -2.7139022   7.10780490 OTU_341                65    FALSE           65
#> 66   -2.8981330   2.37865081 OTU_343                66    FALSE           66
#> 67   -9.4833170  -5.22832057 OTU_346                67    FALSE           67
#> 68   -9.7846769   3.21694191 OTU_386                68    FALSE           68
#> 69  -12.1567051  -5.03251290 OTU_387                69    FALSE           69
#> 70   -2.8406917   5.19881199 OTU_391                70    FALSE           70
#> 71   -6.9164532  -8.20090273 OTU_393                71    FALSE           71
#> 72   -7.9746760 -11.36156893 OTU_394                72    FALSE           72
#> 73  -10.2697090  -7.65501634 OTU_399                73    FALSE           73
#> 74   -4.7814264   2.61780988 OTU_403                74    FALSE           74
#> 75   -4.7674333 -10.78160034 OTU_405                75    FALSE           75
#> 76   -1.6285509   0.06432597 OTU_406                76    FALSE           76
#> 77   -3.5678072 -13.39188343 OTU_408                77    FALSE           77
#> 78   -3.9524019  -0.32103098 OTU_412                78    FALSE           78
#> 79    9.5418395   0.62632246 OTU_417                79    FALSE           79
#> 80   -1.5644195   7.06992692 OTU_422                80    FALSE           80
#> 81   -8.9329129 -10.53825239 OTU_428                81    FALSE           81
#> 82   -1.1904549   1.09454112 OTU_430                82    FALSE           82
#> 83   -2.4779722   3.29410215 OTU_437                83    FALSE           83
#> 84   -6.1461771 -12.57424183 OTU_441                84    FALSE           84
#> 85    0.5212858  -5.25019411 OTU_444                85    FALSE           85
#> 86   -7.1319161   3.33665347 OTU_449                86    FALSE           86
#> 87   -9.3968408   2.38751499 OTU_450                87    FALSE           87
#> 88  -12.8376623  -2.19816066 OTU_461                88    FALSE           88
#> 89    1.8782237   6.23298181 OTU_465                89    FALSE           89
#> 90   -4.2800920   3.66713017 OTU_467                90    FALSE           90
#> 91    0.5963487   1.76664897 OTU_468                91    FALSE           91
#> 92   -4.5189523 -11.98163861 OTU_474                92    FALSE           92
#> 93   -6.8446620 -10.44988536 OTU_476                93    FALSE           93
#> 94    6.0206721   7.72667816 OTU_478                94    FALSE           94
#> 95   -8.3421143   4.87705003 OTU_480                95    FALSE           95
#> 96   -2.0615813   1.74320667 OTU_483                96    FALSE           96
#> 97   -6.2055044  -5.69360677 OTU_494                97    FALSE           97
#> 98    4.1490026 -10.80704499 OTU_499                98    FALSE           98
#> 99   -1.6503278  -9.78919764 OTU_502                99    FALSE           99
#> 100   6.7734765  -8.88386874 OTU_505               100    FALSE          100
#> 101   5.8856148 -10.03241787 OTU_509               101    FALSE          101
#> 102   5.5379128   0.81826082 OTU_512               102    FALSE          102
#> 103  -8.2716324  -7.37245773 OTU_517               103    FALSE          103
#> 104   9.6979800  -5.50518309 OTU_519               104    FALSE          104
#> 105   8.8509865  -1.91774368 OTU_522               105    FALSE          105
#> 106 -10.6965629  -6.76375963 OTU_528               106    FALSE          106
#> 107  10.5145838  -3.40935793 OTU_530               107    FALSE          107
#> 108  10.3529499  -1.05885265 OTU_532               108    FALSE          108
#> 109   3.3168029  -7.06497659 OTU_533               109    FALSE          109
#> 110  10.5973161   1.66480694 OTU_535               110    FALSE          110
#> 111  -7.2878557  -9.29999215 OTU_551               111    FALSE          111
#> 112   9.0650707  -6.49789889 OTU_560               112    FALSE          112
#> 113  -5.3450096  -9.34077049 OTU_563               113    FALSE          113
#> 114   7.1420636   2.56928906 OTU_566               114    FALSE          114
#> 115   7.7285849  -9.81688020 OTU_587               115    FALSE          115
#> 116   9.2978615   1.09176128 OTU_611               116    FALSE          116
#> 117   7.6199311   0.85885319 OTU_625               117    FALSE          117
#> 118  -4.4566158 -13.13549550 OTU_627               118    FALSE          118
#> 119   0.9966670 -13.37546954 OTU_647               119    FALSE          119
#> 120   0.4557394 -11.57779644 OTU_651               120    FALSE          120
#> 121  -4.1168901   1.75815955 OTU_660               121    FALSE          121
#> 122  -3.7043068   3.71598332 OTU_661               122    FALSE          122
#> 123  -4.8239799   4.29946393 OTU_664               123    FALSE          123
#> 124   7.1605858  -2.97105025 OTU_726               124    FALSE          124
#> 125   7.2966180  -5.19579455 OTU_729               125    FALSE          125
#> 126  -0.5578847   0.82535028 OTU_739               126    FALSE          126
#> 127  -2.3625840  -0.07983542 OTU_740               127    FALSE          127
#> 128  10.6470368  -2.43727709 OTU_755               128    FALSE          128
#> 129   4.9448881 -10.96351095 OTU_781               129    FALSE          129
#> 130  -3.7406812   2.56183892 OTU_796               130    FALSE          130
#> 131  -7.3974107   4.01093906 OTU_842               131    FALSE          131
#> 132   3.6507366  -2.66705751 OTU_862               132    FALSE          132
#> 133 -11.6360779  -6.03830428 OTU_864               133    FALSE          133
#> 134  10.0072687  -0.81766510 OTU_874               134    FALSE          134
#> 135   5.0056015   8.24135379 OTU_878               135    FALSE          135
#> 136   2.5915290  -8.36518165 OTU_880               136    FALSE          136
#> 137  -5.6420768   2.71336333 OTU_888               137    FALSE          137
#> 138  -1.1573107   2.76430411 OTU_891               138    FALSE          138
#> 139  -1.6452232   2.59375291 OTU_894               139    FALSE          139
#> 140  -5.4825163  -1.13902312 OTU_895               140    FALSE          140
#> 141  -3.5696502 -10.71013463 OTU_901               141    FALSE          141
#> 142   4.4516676  -5.53955765 OTU_902               142    FALSE          142
#> 143   9.8659812   2.88266624 OTU_904               143    FALSE          143
#> 144   9.5731858  -4.16775008 OTU_906               144    FALSE          144
#> 145  -5.6775491 -11.44520024 OTU_916               145    FALSE          145
#> 146   5.1957234  -8.52168481 OTU_919               146    FALSE          146
#> 147   0.1058623   3.88908846   OTU_9               147    FALSE          147
#> 148  10.7107582  -0.05542644  OTU_11               148    FALSE          148
#> 149   8.6774425   4.44123001  OTU_28               149    FALSE          149
#> 150   4.0375575  -7.72250135  OTU_40               150    FALSE          150
#> 151 -12.2497796  -2.64221410  OTU_58               151    FALSE          151
#> 152 -11.0786869  -4.34391905  OTU_97               152    FALSE          152
#> 153  -0.5747739   6.49865909 OTU_126               153    FALSE          153
#> 154  -4.2436889  -6.95915673 OTU_175               154    FALSE          154
#> 155   1.1722414   7.85131383 OTU_215               155    FALSE          155
#> 156  -4.1704416   2.53507455 OTU_221               156    FALSE          156
#> 157  -6.1388792   4.52817216 OTU_243               157    FALSE          157
#> 158 -12.6384561  -0.76072522 OTU_244               158    FALSE          158
#> 159  -4.2373059   4.70696960 OTU_246               159    FALSE          159
#> 160  -3.3606454   7.26337577 OTU_247               160    FALSE          160
#> 161   2.5130987 -13.03691383 OTU_249               161    FALSE          161
#> 162  -6.3109983   3.33181518 OTU_251               162    FALSE          162
#> 163 -12.0975206  -5.29166500 OTU_252               163    FALSE          163
#> 164  -2.2914013 -13.61199508 OTU_256               164    FALSE          164
#> 165  -7.3016571   1.61740863 OTU_259               165    FALSE          165
#> 166 -13.1717620  -3.85967137 OTU_260               166    FALSE          166
#> 167   8.9613695   1.50839943 OTU_261               167    FALSE          167
#> 168  -7.8083866   2.98692046 OTU_289               168    FALSE          168
#> 169  -4.2933502   6.62749204 OTU_294               169    FALSE          169
#> 170  -5.3764599   5.66213525 OTU_295               170    FALSE          170
#> 171 -10.1726203  -0.81648881 OTU_317               171    FALSE          171
#> 172  -1.8949627  -6.93875267 OTU_336               172    FALSE          172
#> 173   5.9144979   7.72419159 OTU_344               173    FALSE          173
#> 174 -11.0594190  -3.27734547 OTU_396               174    FALSE          174
#> 175  -2.3004428   4.67553113 OTU_400               175    FALSE          175
#> 176  -9.6212866  -8.74793222 OTU_419               176    FALSE          176
#> 177  -8.2641946 -10.08289474 OTU_432               177    FALSE          177
#> 178  -4.1537024 -12.71646785 OTU_457               178    FALSE          178
#> 179  -7.0617321 -11.19582271 OTU_487               179    FALSE          179
#> 180  -9.8129091  -9.23749201 OTU_506               180    FALSE          180
#> 181   1.8032218  -9.73743712 OTU_516               181    FALSE          181
#> 182   8.7586592  -7.91086283 OTU_534               182    FALSE          182
#> 183  -7.4348932   6.10488838 OTU_538               183    FALSE          183
#> 184  -3.6902834   6.17041752 OTU_545               184    FALSE          184
#> 185  -3.1742370   3.09301705 OTU_555               185    FALSE          185
#> 186  -6.4783462   6.01713675 OTU_556               186    FALSE          186
#> 187  -5.3111227   8.13206404 OTU_559               187    FALSE          187
#> 188  -4.4464367   5.63523104 OTU_571               188    FALSE          188
#> 189  -2.3923591   5.78187827 OTU_575               189    FALSE          189
#> 190   6.5452924   5.66567132 OTU_588               190    FALSE          190
#> 191  -7.6940483   2.18973993 OTU_600               191    FALSE          191
#> 192 -11.0923701   0.21680862 OTU_601               192    FALSE          192
#> 193  -3.3832019   4.76597854 OTU_602               193    FALSE          193
#> 194   7.9910851   5.64112046 OTU_604               194    FALSE          194
#> 195  -2.0586968 -12.37620528 OTU_621               195    FALSE          195
#> 196  -5.7564164   7.08601376 OTU_648               196    FALSE          196
#> 197  -4.8358774   3.07222675 OTU_649               197    FALSE          197
#> 198  -3.9663659   7.36877461 OTU_653               198    FALSE          198
#> 199  -2.5180658   3.81949027 OTU_655               199    FALSE          199
#> 200   3.8011768 -10.38886271 OTU_657               200    FALSE          200
#> 201  -7.6827971   4.76035229 OTU_658               201    FALSE          201
#> 202  -1.5246772   8.84991087 OTU_659               202    FALSE          202
#> 203  -4.7635917   6.90890819 OTU_663               203    FALSE          203
#> 204 -11.9000482  -2.51017043 OTU_673               204    FALSE          204
#> 205  -1.7367982 -11.12486841 OTU_680               205    FALSE          205
#> 206  -6.9690222  -7.04832609 OTU_683               206    FALSE          206
#> 207   7.7073745   3.61284811 OTU_702               207    FALSE          207
#> 208  -3.4212023   4.69229491 OTU_704               208    FALSE          208
#> 209   7.1682576   5.34852533 OTU_712               209    FALSE          209
#> 210  -1.1860788   4.19732616 OTU_746               210    FALSE          210
#> 211 -10.5013861  -8.94711817 OTU_793               211    FALSE          211
#> 212  -3.3487573  -7.96640136 OTU_794               212    FALSE          212
#> 213   0.2073409 -13.45034061 OTU_823               213    FALSE          213
#> 214  -4.8823099   0.23090176 OTU_830               214    FALSE          214
#> 215   0.2894501 -12.03537455 OTU_834               215    FALSE          215
#> 216  -9.8920492  -5.80611498 OTU_836               216    FALSE          216
#> 217 -10.7760748  -4.02162713 OTU_844               217    FALSE          217
#> 218  -6.6389331   6.66485515 OTU_872               218    FALSE          218
#> 219  -3.7567489  -9.93955487 OTU_883               219    FALSE          219
#> 220   9.2558176   2.88608903 OTU_892               220    FALSE          220