Compute proportions for taxa.

Computes the proportion of a taxa classification. Function from the phylosmith-package.

taxa_proportions(phyloseq_obj, classification, treatment = NA)

Arguments

phyloseq_obj
A phyloseq-class object. it must contain tax_table()) with information about each taxa/gene.

classification
Column name as a string or numeric in the tax_table for the proportions to be reported on.

treatment
Column name as a string or numeric in the sample_data. This can be a vector of multiple columns and they will be combined into a new column. If NA, then proportions will be reported for the entire dataset. If set to "sample" it will report proportions by sample.

Value

data.table

Examples

taxa_proportions(soil_column, 'Phylum', treatment = NA)

#>                          Phylum Proportion
#>  1:              Proteobacteria      0.245
#>  2:               Acidobacteria      0.162
#>  3:             Verrucomicrobia      0.114
#>  4:       unclassified_Bacteria      0.120
#>  5:              Actinobacteria      0.079
#>  6:                  Firmicutes      0.038
#>  7:               Bacteroidetes      0.055
#>  8:    candidate division WPS-1      0.032
#>  9:                 Chloroflexi      0.052
#> 10:              Planctomycetes      0.052
#> 11:                Spirochaetes      0.004
#> 12:   Cyanobacteria/Chloroplast      0.002
#> 13:            Gemmatimonadetes      0.007
#> 14:              Woesearchaeota      0.002
#> 15:              Thaumarchaeota      0.013
#> 16:             Latescibacteria      0.005
#> 17:               Euryarchaeota      0.005
#> 18:               Parcubacteria      0.002
#> 19:           unclassified_Root      0.001
#> 20:             Armatimonadetes      0.007
#> 21:               Fibrobacteres      0.000
#> 22:                  Chlamydiae      0.001
#> 23:                 Nitrospirae      0.002
#> 24:    candidate division WPS-2      0.000
#> 25:               Pacearchaeota      0.000
#> 26:                 Tenericutes      0.000
#> 27:               Synergistetes      0.000
#> 28:               Crenarchaeota      0.000
#> 29:        unclassified_Archaea      0.000
#> 30:                        BRC1      0.001
#> 31: Candidatus Saccharibacteria      0.000
#> 32:         Deinococcus-Thermus      0.000
#> 33:              Microgenomates      0.000
#> 34:             Hydrogenedentes      0.000
#> 35:               Cloacimonetes      0.000
#> 36:             Ignavibacteriae      0.000
#> 37:              Diapherotrites      0.000
#> 38:               Lentisphaerae      0.000
#> 39:      candidate division ZB3      0.000
#> 40:                         SR1      0.000
#> 41:               Elusimicrobia      0.000
#> 42:                Fusobacteria      0.000
#> 43:             Deferribacteres      0.000
#>                          Phylum Proportion

taxa_proportions(soil_column, 'Phylum', treatment = 'sample')

#>        Sample                   Phylum Proportion
#>     1:    S_0               Firmicutes      0.637
#>     2:    S_0            Acidobacteria      0.071
#>     3:    S_0            Euryarchaeota      0.061
#>     4:    S_0           Actinobacteria      0.043
#>     5:    S_0           Proteobacteria      0.037
#>    ---                                           
#> 20679:   S_99            Parcubacteria      0.000
#> 20680:   S_99            Synergistetes      0.000
#> 20681:   S_99 candidate division WPS-2      0.000
#> 20682:   S_99   candidate division ZB3      0.000
#> 20683:   S_99     unclassified_Archaea      0.000

taxa_proportions(soil_column, 'Phylum', treatment = c('Matrix', 'Treatment'))

#>        Matrix_Treatment         Phylum Proportion
#>   1:     Manure Control     Firmicutes      0.627
#>   2:     Manure Control  Acidobacteria      0.069
#>   3:     Manure Control  Euryarchaeota      0.054
#>   4:     Manure Control Proteobacteria      0.041
#>   5:     Manure Control  Bacteroidetes      0.027
#>  ---                                             
#> 206: Effluent Unamended    Tenericutes      0.000
#> 207: Effluent Unamended   Fusobacteria      0.000
#> 208: Effluent Unamended  Elusimicrobia      0.000
#> 209: Effluent Unamended  Fibrobacteres      0.000
#> 210: Effluent Unamended  Lentisphaerae      0.000